Don't make the arraysize() macro cast to int.
Reviewed-on: https://code-review.googlesource.com/c/re2/+/43030
(cherry picked from commit 7e9d7987aabfd14198f1380de8e0057d8992ac05)

+++ re2/re2.cc
@@ -368,7 +368,7 @@ bool RE2::Replace(std::string* str,
                   absl::string_view rewrite) {
   absl::string_view vec[kVecSize];
   int nvec = 1 + MaxSubmatch(rewrite);
-  if (nvec > arraysize(vec))
+  if (nvec > static_cast<int>(arraysize(vec)))
     return false;
   if (!re.Match(*str, 0, str->size(), UNANCHORED, vec, nvec))
     return false;
@@ -388,7 +388,7 @@ int RE2::GlobalReplace(std::string* str,
                        absl::string_view rewrite) {
   absl::string_view vec[kVecSize];
   int nvec = 1 + MaxSubmatch(rewrite);
-  if (nvec > arraysize(vec))
+  if (nvec > static_cast<int>(arraysize(vec)))
     return false;
 
   const char* p = str->data();
@@ -460,7 +460,7 @@ bool RE2::Extract(absl::string_view text,
                   std::string* out) {
   absl::string_view vec[kVecSize];
   int nvec = 1 + MaxSubmatch(rewrite);
-  if (nvec > arraysize(vec))
+  if (nvec > static_cast<int>(arraysize(vec)))
     return false;
 
   if (!re.Match(text, 0, text.size(), UNANCHORED, vec, nvec))
